test: move class in TestCase file
$ vendor/bin/phpunit tests/system/Test/TestCaseTest.php
PHPUnit 9.6.15 by Sebastian Bergmann and contributors.

Runtime:       PHP 8.1.26
Configuration: /.../CodeIgniter4/phpunit.xml

E........                                                                                                 9 / 9 (100%)

Time: 00:00.086, Memory: 16.00 MB

There was 1 error:

1) CodeIgniter\Test\TestCaseTest::testGetPrivatePropertyWithObject
Error: Class "CodeIgniter\Test\__TestForReflectionHelper" not found

/.../CodeIgniter4/tests/system/Test/TestCaseTest.php:38

ERRORS!
Tests: 9, Assertions: 19, Errors: 1.
