+++ tests/system/Test/ReflectionHelperTest.php
@@ -11,6 +11,8 @@
 
 namespace CodeIgniter\Test;
 
+use Tests\Support\Test\TestForReflectionHelper;
+
 /**
  * @internal
  *
@@ -20,14 +22,14 @@ final class ReflectionHelperTest extends CIUnitTestCase
 {
     public function testGetPrivatePropertyWithObject(): void
     {
-        $obj    = new __TestForReflectionHelper();
+        $obj    = new TestForReflectionHelper();
         $actual = $this->getPrivateProperty($obj, 'private');
         $this->assertSame('secret', $actual);
     }
 
     public function testGetPrivatePropertyWithObjectStaticCall(): void
     {
-        $obj    = new __TestForReflectionHelper();
+        $obj    = new TestForReflectionHelper();
         $actual = CIUnitTestCase::getPrivateProperty($obj, 'private');
         $this->assertSame('secret', $actual);
     }
@@ -35,7 +37,7 @@ final class ReflectionHelperTest extends CIUnitTestCase
     public function testGetPrivatePropertyWithStatic(): void
     {
         $actual = $this->getPrivateProperty(
-            __TestForReflectionHelper::class,
+            TestForReflectionHelper::class,
             'static_private'
         );
         $this->assertSame('xyz', $actual);
@@ -43,7 +45,7 @@ final class ReflectionHelperTest extends CIUnitTestCase
 
     public function testSetPrivatePropertyWithObject(): void
     {
-        $obj = new __TestForReflectionHelper();
+        $obj = new TestForReflectionHelper();
         $this->setPrivateProperty(
             $obj,
             'private',
@@ -55,19 +57,19 @@ final class ReflectionHelperTest extends CIUnitTestCase
     public function testSetPrivatePropertyWithStatic(): void
     {
         $this->setPrivateProperty(
-            __TestForReflectionHelper::class,
+            TestForReflectionHelper::class,
             'static_private',
             'abc'
         );
         $this->assertSame(
             'abc',
-            __TestForReflectionHelper::getStaticPrivate()
+            TestForReflectionHelper::getStaticPrivate()
         );
     }
 
     public function testGetPrivateMethodInvokerWithObject(): void
     {
-        $obj    = new __TestForReflectionHelper();
+        $obj    = new TestForReflectionHelper();
         $method = $this->getPrivateMethodInvoker(
             $obj,
             'privateMethod'
@@ -81,38 +83,12 @@ final class ReflectionHelperTest extends CIUnitTestCase
     public function testGetPrivateMethodInvokerWithStatic(): void
     {
         $method = $this->getPrivateMethodInvoker(
-            __TestForReflectionHelper::class,
+            TestForReflectionHelper::class,
             'privateStaticMethod'
         );
         $this->assertSame(
             'private_static param1param2',
             $method('param1', 'param2')
         );
-    }
-}
-
-class __TestForReflectionHelper
-{
-    private string $private               = 'secret';
-    private static string $static_private = 'xyz';
-
-    public function getPrivate()
-    {
-        return $this->private;
-    }
-
-    public static function getStaticPrivate()
-    {
-        return self::$static_private;
-    }
-
-    private function privateMethod($param1, $param2)
-    {
-        return 'private ' . $param1 . $param2;
-    }
-
-    private static function privateStaticMethod($param1, $param2)
-    {
-        return 'private_static ' . $param1 . $param2;
     }
 }
